pyproject.toml
[project]
name = "langgraph-rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langgraph>=0.5.3",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "streamlit>=1.46.1",
]
